merging

+++ src/main/java/com/techelevator/Chip.java
@@ -1,10 +1,12 @@
 package com.techelevator;
 
+import java.math.BigDecimal;
+
 public class Chip extends Product {
 
     private String chipSlogan = "Crunch Crunch";
 
-    public Chip(String productName, double productPrice) {
+    public Chip(String productName, BigDecimal productPrice) {
         super(productName, productPrice);
 
     }
